# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ColorCount {
  name: string;
  code: string;
  hex: string;
  count: number;
  beforeCleanup: number;
}

export interface SectionInfo {
  label: string;
  row: number;
  col: number;
  filename: string;
}

export interface MosaicResult {
  sessionId: string;
  totalStuds: number;
  width: number;
  height: number;
  baseplateSize: number;
  columns: number;
  rows: number;
  colorsBefore: number;
  colorsAfter: number;
  colorCounts: ColorCount[];
  sections: SectionInfo[];
  previewUrl: string;
  layoutUrl: string;
  csvUrl: string;
  zipUrl: string;
}

export interface ErrorResponse {
  error: string;
}

/**
 * Detail mode
 */
export type GenerateMosaicBodyMode =
  (typeof GenerateMosaicBodyMode)[keyof typeof GenerateMosaicBodyMode];

export const GenerateMosaicBodyMode = {
  detail: "detail",
  balanced: "balanced",
  clean: "clean",
} as const;

export type GenerateMosaicBodyPaletteItemFamily =
  (typeof GenerateMosaicBodyPaletteItemFamily)[keyof typeof GenerateMosaicBodyPaletteItemFamily];

export const GenerateMosaicBodyPaletteItemFamily = {
  neutral: "neutral",
  skin: "skin",
  brown: "brown",
  red: "red",
  orange: "orange",
  yellow: "yellow",
  green: "green",
  blue: "blue",
  cyan: "cyan",
  purple: "purple",
  pink: "pink",
  metallic: "metallic",
} as const;

export type GenerateMosaicBodyPaletteItem = {
  name: string;
  code: string;
  hex: string;
  enabled: boolean;
  family: GenerateMosaicBodyPaletteItemFamily;
};

export type GenerateMosaicBody = {
  image: Blob;
  /** Baseplate size (16 or 32) */
  baseplateSize: number;
  /** Number of baseplate columns */
  columns: number;
  /** Number of baseplate rows */
  rows: number;
  /** Detail mode */
  mode?: GenerateMosaicBodyMode;
  /** Protect high-contrast edges during cleanup */
  protectEdges?: boolean;
  /** Full palette sent from frontend */
  palette?: GenerateMosaicBodyPaletteItem[];
};