# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: API specification

servers:
  - url: /api
    description: Base API path

tags:
  - name: health
    description: Health operations
  - name: mosaic
    description: Mosaic generation operations

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /mosaic/generate:
    post:
      operationId: generateMosaic
      tags: [mosaic]
      summary: Generate a LEGO mosaic from an uploaded image
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - image
                - baseplateSize
                - columns
                - rows
              properties:
                image:
                  type: string
                  format: binary
                baseplateSize:
                  type: integer
                  description: Baseplate size (16 or 32)
                columns:
                  type: integer
                  description: Number of baseplate columns
                rows:
                  type: integer
                  description: Number of baseplate rows

                mode:
                  type: string
                  description: Detail mode
                  enum:
                    - detail
                    - balanced
                    - clean
                  default: balanced

                protectEdges:
                  type: boolean
                  description: Protect high-contrast edges during cleanup
                  default: true

                palette:
                  type: array
                  description: Full palette sent from frontend
                  items:
                    type: object
                    properties:
                      name:
                        type: string
                      code:
                        type: string
                      hex:
                        type: string
                      enabled:
                        type: boolean
                      family:
                        type: string
                        enum:
                          - neutral
                          - skin
                          - brown
                          - red
                          - orange
                          - yellow
                          - green
                          - blue
                          - cyan
                          - purple
                          - pink
                          - metallic
                    required:
                      - name
                      - code
                      - hex
                      - enabled
                      - family

      responses:
        "200":
          description: Mosaic generated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MosaicResult"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /mosaic/download/{sessionId}/{filename}:
    get:
      operationId: downloadFile
      tags: [mosaic]
      summary: Download a generated mosaic file
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
        - name: filename
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: File download
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        "404":
          description: File not found

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ColorCount:
      type: object
      properties:
        name:
          type: string
        code:
          type: string
        hex:
          type: string
        count:
          type: integer
        beforeCleanup:
          type: integer
      required:
        - name
        - code
        - hex
        - count
        - beforeCleanup

    SectionInfo:
      type: object
      properties:
        label:
          type: string
        row:
          type: integer
        col:
          type: integer
        filename:
          type: string
      required:
        - label
        - row
        - col
        - filename

    MosaicResult:
      type: object
      properties:
        sessionId:
          type: string
        totalStuds:
          type: integer
        width:
          type: integer
        height:
          type: integer
        baseplateSize:
          type: integer
        columns:
          type: integer
        rows:
          type: integer
        colorsBefore:
          type: integer
        colorsAfter:
          type: integer
        colorCounts:
          type: array
          items:
            $ref: "#/components/schemas/ColorCount"
        sections:
          type: array
          items:
            $ref: "#/components/schemas/SectionInfo"
        previewUrl:
          type: string
        layoutUrl:
          type: string
        csvUrl:
          type: string
        zipUrl:
          type: string
      required:
        - sessionId
        - totalStuds
        - width
        - height
        - baseplateSize
        - columns
        - rows
        - colorsBefore
        - colorsAfter
        - colorCounts
        - sections
        - previewUrl
        - layoutUrl
        - csvUrl
        - zipUrl

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error