# lib/api-zod/src/generated/types/colorCount.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ColorCount {
  name: string;
  code: string;
  hex: string;
  count: number;
  beforeCleanup: number;
}