# lib/api-zod/src/generated/types/mosaicResult.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ColorCount } from "./colorCount";
import type { SectionInfo } from "./sectionInfo";

export interface MosaicResult {
  sessionId: string;
  totalStuds: number;
  width: number;
  height: number;
  baseplateSize: number;
  columns: number;
  rows: number;
  colorsBefore: number;
  colorsAfter: number;
  colorCounts: ColorCount[];
  sections: SectionInfo[];
  previewUrl: string;
  layoutUrl: string;
  csvUrl: string;
  zipUrl: string;
}