# lib/api-zod/src/generated/types/sectionInfo.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SectionInfo {
  label: string;
  row: number;
  col: number;
  filename: string;
}